+++ apps/native-component-list/src/navigation/ExpoComponentsStackNavigator.tsx
@@ -11,6 +11,7 @@ import { Layout } from '../constants';
 import { CameraScreens } from '../screens/Camera/CameraScreen';
 import ExpoComponents from '../screens/ExpoComponentsScreen';
 import { ImageScreens } from '../screens/Image/ImageScreen';
+import { UIScreens } from '../screens/UI/UIScreen';
 import { VideoScreens } from '../screens/Video/VideoScreen';
 import { ScreenConfig } from '../types/ScreenConfig';
 
@@ -378,7 +379,7 @@ export const Screens: ScreenConfig[] = [
   },
   {
     getComponent() {
-      return optionalRequire(() => require('../screens/UIScreen'));
+      return optionalRequire(() => require('../screens/UI/UIScreen'));
     },
     name: 'Expo UI',
   },
@@ -434,6 +435,7 @@ export const Screens: ScreenConfig[] = [
   ...CameraScreens,
   ...ImageScreens,
   ...VideoScreens,
+  ...UIScreens,
 ];
 
 function ExpoComponentsStackNavigator(props: { navigation: BottomTabNavigationProp<any> }) {